rename NameTag => Name
