+++ app/components/UI/Name/Name.types.ts
@@ -1,0 +1,3 @@
+export interface NameProperties {
+  address: string;
+}
