--- app/components/UI/NameTag/NameTag.types.ts
@@ -1,3 +1,0 @@
-export interface NameTagProperties {
-  address: string;
-}
