--- app/components/UI/NameTag/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NameTag';
