+++ storybook/storyLoader.js
@@ -68,7 +68,7 @@ function loadStories() {
   require('../app/components/Base/Title.stories');
   require('../app/components/UI/ButtonReveal/ButtonReveal.stories');
   require('../app/components/UI/Fox/Fox.stories');
-  require('../app/components/UI/NameTag/NameTag.stories');
+  require('../app/components/UI/Name/Name.stories');
   require('../app/components/UI/StyledButton/StyledButton.stories');
   require('../app/components/UI/TextEstimated/TextEstimated.stories');
   require('../app/components/UI/WarningAlert/WarningAlert.stories');
@@ -144,7 +144,7 @@ const stories = [
   '../app/components/Base/Title.stories',
   '../app/components/UI/ButtonReveal/ButtonReveal.stories',
   '../app/components/UI/Fox/Fox.stories',
-  '../app/components/UI/NameTag/NameTag.stories',
+  '../app/components/UI/Name/Name.stories',
   '../app/components/UI/StyledButton/StyledButton.stories',
   '../app/components/UI/TextEstimated/TextEstimated.stories',
   '../app/components/UI/WarningAlert/WarningAlert.stories',
